requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
markdown==3.5.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
